Got dropdown refs working, starting working on doc info popup, the logo svg broke completely don't know what thats about
