+++ src/index.css
@@ -61,3 +61,12 @@ code {
     scale: 1;
   }
 }
+
+@keyframes fade-in {
+  0% {
+    opacity: 0;
+  };
+  100% {
+    opacity: 1;
+  }
+}
