+++ src/scripts/clearText.js
@@ -1,0 +1,5 @@
+export const clearText = () => {
+    const textfield = document.getElementById("textfield");
+    const titlefield = document.getElementById("titlefield");
+    textfield.value.length != 0 ? (textfield.value = "") : (titlefield.value = "");
+};
