+++ src/scripts/copyText.js
@@ -1,0 +1,5 @@
+export const copyText = () => {
+    const textfield = document.getElementById("textfield");
+    textfield.select();
+    document.execCommand("copy");
+};
